fix: double to float

+++ imu_embedding.py
@@ -17,7 +17,7 @@ def get_embeddings(extrapolated_imu_data, device):
 
     # Loop through data points and sensors
     for i in range(N):
-        time_series = extrapolated_imu_data[i,:,:6]
+        time_series = extrapolated_imu_data[i,:,:6].astype(np.float32)
         time_series = np.expand_dims(np.transpose(time_series), axis=0)
         print(time_series.shape)
         # Load data
